Update to SharpGenTools.Sdk

+++ CorApi2/debug/Constants.cs
@@ -75,17 +75,17 @@ namespace Microsoft.Samples.Debugging.CorDebug
 
     public abstract class TokenUtils
     {
-        public static CorTokenType TypeFromToken(int token)
+        public static CorTokenType TypeFromToken(uint token)
         {
-            return (CorTokenType) ((UInt32)token & 0xff000000);
+            return (CorTokenType) (token & 0xff000000);
         }
 
-        public static int RidFromToken(int token)
+        public static int RidFromToken(uint token)
         {
-            return (int)( (UInt32)token & 0x00ffffff);
+            return (int)(token & 0x00ffffff);
         }
 
-        public static bool IsNullToken(int token)
+        public static bool IsNullToken(uint token)
         {
             return (RidFromToken(token)==0);
         }
